Continue setup (#4)
* Add launch.json

Add the ability to debug BFF via docker using VSCode.

* Add settings.json

set the file endings

* Update Dockerfile.dev

Update the docker file to match that of the identity and api service.
Add entrypoint.sh to add certs to CA and kick off dotnet watch command

* Wire up OIDC and Cookie Auth

User/Login triggers the OIDC workflow to redirect the user to IDP.
Tokens are captured in a cookie
Cookie auth make the claims available to the pipeline.

* Update launch.json

Update the name, so it's easier to tell the difference.

* Update appsettings.Development.json

Update Authority to match issuer.

* Update AsgOpenIdConnectEvents.cs

IssuerAddress does not need to be set.

* remove blank line

* Add RedirectMiddleware

When running in development environment, redirect middleware inspects the response and if it's one of the redirects specified via environment variables. the path is updated.