Update AsgOpenIdConnectEvents.cs
Update OnTokenValidated method to capture the tokens after they have been fetched. No reason to validate them, since they were fetched directly from IDP.
The tokens are cached using InMemoryCache. The key is the value of the Sub claim in the access token.
The value of the Sub claim is stored in a cookie.