Add AsgJwtBearerEvents.cs
Provide an event handler for OnMessageReceived event.
The event is invoked when JWT Authentication is being invoked.
The event handler checks for the cookie created by AsgOpenIdConnectEvents.OnTokenValidated event.
If the cookie exists, it means the user has authenticated successfully. If the cookie does not exist, the user has not authenticated.
If the cookie exists, the sub claim value is fetched from the cookie and Tokens for the users are retrieved from the cache.
Claims principal is created and access token is stored in the HttpContext. So GetTokenAsync extennsion method on the HttpContext will be able to retrieve the token if requested.