Save tokens to cache (#5)
* Update AsgOpenIdConnectEvents.cs

removed unnecessary methods

* Remove TestApiController.cs

Controller was only created to help with wiring up of OIDC and JWT Authentications

* Add UserAuthenticationInformation.cs

Model represents user's tokens.

* Update AsgOpenIdConnectEvents.cs

Update OnTokenValidated method to capture the tokens after they have been fetched. No reason to validate them, since they were fetched directly from IDP.
The tokens are cached using InMemoryCache. The key is the value of the Sub claim in the access token.
The value of the Sub claim is stored in a cookie.

* Update simple-budget.bff.csproj

Add Microsoft.AspNetCore.Authentication.JwtBearer package.

* Add AsgJwtBearerEvents.cs

Provide an event handler for OnMessageReceived event.
The event is invoked when JWT Authentication is being invoked.
The event handler checks for the cookie created by AsgOpenIdConnectEvents.OnTokenValidated event.
If the cookie exists, it means the user has authenticated successfully. If the cookie does not exist, the user has not authenticated.
If the cookie exists, the sub claim value is fetched from the cookie and Tokens for the users are retrieved from the cache.
Claims principal is created and access token is stored in the HttpContext. So GetTokenAsync extennsion method on the HttpContext will be able to retrieve the token if requested.

* Add CookieEvents.cs

Add event handlers for OnRedirectToLogin and OnRedirectToAccessDenied,
instead redirects, status codes are returned.

* Update UserController.cs

Update Login method to check if user is Authenticated.
If the user is Authenticated, redirect to LoggedIn page.
Otherwise, Challenge the request. This triggers OIDC login flow. The user will be redirect to Identity Provider.

* Update HostingExtensions.cs

Configure Authentication schemes.
DefaultSignInScheme is Cookie.
DefaultAuthenticateScheme is JWTBearer
DefaultChallengeScheme is OIDC
Update the CookieAuth Options to forward the Challenge, Authenticate requests to JWT and OIDC respectively and wire up the OnRedirectToLogin
and OnRedirectToAccessDenied event-handlers.
Wire up JwtBearer Authentication with OnMessageRecieved event-handler.
Add InMemoryCache.