Add CookieService.cs
Add ICookieService interface and CookieService Implementation. Allows for setting and reading from a cookie.
The CookieService also uses the DataProtection APIs to encrpyt the data in cookie.
Update AsgJwtBearerEvents and AsgOpenIdConnectEvents to use the cookie service to set and read from the cookie.
Update scopes being requested, Otherwise email and id claims are not avilable to the API.